+++ tickets/static/tickets/css/slider.css
@@ -28,6 +28,7 @@ li {
 	width: 1000px;
 	height: 300px;
 	float: left;
+	cursor: pointer;
 }
 
 #slider-wrapper ul li:nth-child(1) {
