+++ tickets/static/tickets/js/simulateDataBase.js
@@ -1,0 +1,71 @@
+const btnLive = 'Contact a sales rep';
+const btnSource = 'Buy Tickets';
+const gitIcon = '../../general/img/source-github-btn.svg';
+const buyIcon = '../../general/img/source-live-btn.svg'
+
+const works = [
+  {
+    name: 'Event Name 1',
+    cat: ['Teatro', 'Musical', 2023],
+    description:
+      `Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s. Lorem Ipsum is simply dummy text of the printing and typesetting industry. 
+      Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum han printer took a galley of type and scrambled it 1960s with the releawn printer took a galley of type and scrambled it 1960s.`,
+    tags: ['Teatro', 'Musical', 'JavaScript'],
+    snapshot: '../../general/img/leaves01.jpg',
+    connectionsBtn: [gitIcon, buyIcon],
+    connectionstext: [btnLive, btnSource],
+    connectionsURL: ['https://lordksix.github.io/portafolio-app/', 'https://github.com/lordksix/portafolio-app'],
+  },
+  {
+    name: 'Event Name 2',
+    cat: ['Teatro', 'Drama', 2023],
+    description:
+    `Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s. Lorem Ipsum is simply dummy text of the printing and typesetting industry. 
+    Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum han printer took a galley of type and scrambled it 1960s with the releawn printer took a galley of type and scrambled it 1960s.`,
+    tags: ['HTML', 'CSS', 'JavaScript'],
+    snapshot: '../../general/img/leaves02.jpg',
+    connectionsBtn: [gitIcon, buyIcon],
+    connectionstext: [btnLive, btnSource],
+    connectionsURL: ['https://lordksix.github.io/portafolio-app/', 'https://github.com/lordksix/portafolio-app'],
+  },
+  {
+    name: 'Event Name 3',
+    cat: ['Concierto', 'Rock', 2023],
+    description:
+    `Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s. Lorem Ipsum is simply dummy text of the printing and typesetting industry. 
+    Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum han printer took a galley of type and scrambled it 1960s with the releawn printer took a galley of type and scrambled it 1960s.`,
+    tags: ['HTML', 'CSS', 'JavaScript'],
+    snapshot: '../../general/img/leaves03.jpg',
+    connectionsBtn: [gitIcon, buyIcon],
+    connectionstext: [btnLive, btnSource],
+    connectionsURL: ['https://lordksix.github.io/portafolio-app/', 'https://github.com/lordksix/portafolio-app'],
+  },
+  {
+    name: 'Event Name 4',
+    cat: ['Concierto', 'Internacional', 2024],
+    description:
+    `Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s. Lorem Ipsum is simply dummy text of the printing and typesetting industry. 
+    Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum han printer took a galley of type and scrambled it 1960s with the releawn printer took a galley of type and scrambled it 1960s.`,
+    tags: ['HTML', 'CSS', 'Ruby on Rails', 'JavaScript'],
+    snapshot: '../../general/img/leaves04.jpg',
+    connectionsBtn: [gitIcon, buyIcon],
+    connectionstext: [btnLive, btnSource],
+    connectionsURL: ['https://lordksix.github.io/portafolio-app/', 'https://github.com/lordksix/portafolio-app'],
+  },
+  {
+    name: 'Event Name 5',
+    cat: ['Teatro', 'Musical', 2024],
+    description:
+    `Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s. Lorem Ipsum is simply dummy text of the printing and typesetting industry. 
+    Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum han printer took a galley of type and scrambled it 1960s with the releawn printer took a galley of type and scrambled it 1960s.`,
+    tags: ['HTML', 'CSS', 'Ruby on Rails', 'JavaScript'],
+    snapshot: '../../general/img/leaves05.jpg',
+    connectionsBtn: [gitIcon, buyIcon],
+    connectionstext: [btnLive, btnSource],
+    connectionsURL: ['https://lordksix.github.io/portafolio-app/', 'https://github.com/lordksix/portafolio-app'],
+  },
+];
+
+export {
+  works,
+}
