+++ tickets/templates/tickets/main.html
@@ -12,7 +12,7 @@
   <h1>Proximos eventos de la Lorem ipsumboo</h1>
   <div id="slider-wrapper">
     <ul>
-      <li><a href="" data-eventID="0">
+      <li data-eventID="0">
         <div class="event-slide">
           <h2>Slide Event</h2>
           <p>
@@ -21,9 +21,8 @@
             est, placeat facere non vero culpa inventore vitae!
           </p>
         </div>
-      </a>
       </li>
-      <li><a href="" data-eventID="1">
+      <li data-eventID="1">
         <div class="event-slide">
           <h2>Slide Event</h2>
           <p>
@@ -32,9 +31,8 @@
             est, placeat facere non vero culpa inventore vitae!
           </p>
         </div>
-      </a>
       </li>
-      <li><a href="" data-eventID="2">
+      <li data-eventID="2">
         <div class="event-slide">
           <h2>Slide Event</h2>
           <p>
@@ -43,9 +41,8 @@
             est, placeat facere non vero culpa inventore vitae!
           </p>
         </div>
-      </a>
       </li>
-      <li><a href="" data-eventID="3">
+      <li data-eventID="3">
         <div class="event-slide">
           <h2>Slide Event</h2>
           <p>
@@ -54,9 +51,8 @@
             est, placeat facere non vero culpa inventore vitae!
           </p>
         </div>
-      </a>
       </li>
-      <li><a href="" data-eventID="4">
+      <li data-eventID="4">
         <div class="event-slide">
           <h2>Slide Event</h2>
           <p>
@@ -65,7 +61,6 @@
             est, placeat facere non vero culpa inventore vitae!
           </p>
         </div>
-      </a>
       </li>
     </ul>
     <div id="slider-controls">
